requirements.txt
python-telegram-bot
tinydb
python-dotenv
python-docx